+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - New media:fix-nonlocal-driver command. Fixes s3 media created with invalid FILESYSTEM_DRIVER=s3 configuration ([672cccd4](https://github.com/pixelfed/pixelfed/commit/672cccd4))
+- New landing page design ([09c0032b](https://github.com/pixelfed/pixelfed/commit/09c0032b))
 
 ### Updates
 - Update ApiV1Controller, fix blocking remote accounts. Closes #4256 ([8e71e0c0](https://github.com/pixelfed/pixelfed/commit/8e71e0c0))